feat: refactor factory structure and update import paths
